Added PSR-4 autoloading. Cleaned up some of the documentation. Added PSR-2 compliance.
