+++ composer.json
@@ -3,5 +3,13 @@
     "description": "Kerio APIs Client Library for PHP",
     "require": {
         "php": ">=5.3.0"
+    },
+    "autoload": {
+        "psr-4": {
+            "Kerio\\Api\\": "src/"
+        }
+    },
+    "require-dev": {
+        "squizlabs/php_codesniffer": "^3.0"
     }
 }
