+++ src/Http/KerioApiSocketInterface.php
@@ -12,21 +12,23 @@
  * Any changes may be overwritten by a new version.
  */
 
+namespace Kerio\Api\Http;
+
 /**
  * Kerio API Socket Interface.
  *
  * This interface describes basic methods used in HTTP communication.
  *
- * @copyright	Copyright &copy; 2012-2012 Kerio Technologies s.r.o.
- * @license		http://www.kerio.com/developers/license/sdk-agreement
- * @version		1.4.0.234
+ * @copyright Copyright &copy; 2012-2012 Kerio Technologies s.r.o.
+ * @license http://www.kerio.com/developers/license/sdk-agreement
+ * @version 1.4.0.234
  */
-interface KerioApiSocketInterface {
-
-	/**
-	 * Send data to socket.
-	 *
-	 * @param	string
-	 */
-	public function send($data);
+interface KerioApiSocketInterface
+{
+    /**
+     * Send data to socket.
+     *
+     * @param string $data
+     */
+    public function send($data);
 }
